Finished Appointment class, completed a rudimentary PrintInvoice method with an InvoiceFrame inner class. Added toString methods for PatientInfo and Bill to make printing the Invoice easier.
Deleted "Time" attribute of Appointment and added a Patient attribute. In accordance with this, also updated each class that uses the Appointment constructor accordingly. 

Also added a TODO suggesting a "name" attribute for all users as well as the standard UserName.

The commit failed last time I tried to submit changes. Sorry if this got submitted twice.